[refactor]ユーザ情報をコンポーネント化
